[test] refactor: split tests/e2e/test_e2e_parallel.py by modality
The single 417-line file mixed text / VLM / omni / DiT parametrizations
with the shared torchrun+compare harness, so every model onboarding
diff touched one enormous file. Split into four per-modality files
plus a `_harness.py` module that owns `main()`, default tolerances and
the v4/v5/dit skip marks.

- tests/e2e/_harness.py — shared harness (main, _materialize_weights_dir,
  DEFAULT_RTOL/ATOL, v4_only / v5_only / dit_only marks).
- tests/e2e/test_e2e_parallel_text.py — text LLMs.
- tests/e2e/test_e2e_parallel_vlm.py — Qwen2-VL / Qwen3-VL family +
  Qwen3.5(-MoE) that share the VL dataset fixture.
- tests/e2e/test_e2e_parallel_omni.py — Qwen2.5-Omni / Qwen3-Omni-MoE.
- tests/e2e/test_e2e_parallel_dit.py — Wan DiT. Carries pytestmark =
  [e2e, dit] so the GPU e2e runner can select it with
  `-m "e2e and dit"` in the dedicated `--extra dit` step.

Verified `pytest --collect-only -m e2e tests/e2e` collects the same
22 cases as before, and `-m "e2e and dit"` selects only the Wan case.

Updated all docs and the v5 migration skill to point at the new paths
and import names (`DEFAULT_RTOL`, `v5_only`, etc. from `_harness.py`).
Dataset fixtures already live in tests/e2e/conftest.py so no fixture
relocation was needed.

Made-with: Cursor
