+++ docs/usage/support_new_models/guide_and_checklist.md
@@ -209,7 +209,7 @@ For implementation details of each patch, refer to the example docs.
 - [ ] `DummyYourModelDataset` in `veomni/data/dummy_dataset.py` (multimodal)
 - [ ] `MODEL_TO_DATASET` entry in `tests/models/utils.py`
 - [ ] `pytest.param` in `test_cases` in `tests/models/test_models_patch.py` (Level 1)
-- [ ] Test case + fixture + test function in `tests/e2e/test_e2e_parallel.py` (Level 2)
+- [ ] Test case + (optional new) fixture + test function in `tests/e2e/test_e2e_parallel_{text,vlm,omni,dit}.py` (Level 2) — fixtures live in `tests/e2e/conftest.py`
 - [ ] For VLM models, add the toy config to the `freeze_vit` smoke test list in `tests/models/test_vlm_trainer.py`
 
 ---
